@types 분리
