+++ src/@types/@api/Campagin.d.ts
